+++ lib/time_value_formatter.dart
@@ -131,7 +131,7 @@ class TimeValueFormatter {
       // Curiosities
       case ValueType.binaryClockString:
         return TimeUtils.binaryTimeString(dt);
-      case ValueType.sevenSegmentTime:
+      case ValueType.sevenSegmentClock:
       case ValueType.binaryClockColumns:
       case ValueType.binaryClockBcd:
       default: