+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - 🆒 1-value always-on focus screen with font color choice (and brightness adjust for Android)
+- new value type sevenSegmentClock
 
 ### Changed
 - much improved timezone keyword search